Change database to PostgreSQL (#5)
DuckDB is not suited for scaling web applications as it only supports
one connection at a time. This would make it impossible to dockerize
front-end and spin up multiple servers, or to query/update data while
application is running. DuckDB is more like a replacement for pandas if
it was desired to do local analytics using SQL. Since this is not our
desired use case, replacing with PostgreSQL which is a better fit.